Move MockGenerator to package com.github.jmetzz.mock and enable MockGenertor in bean.xml alternatives
