+++ src/main/resources/META-INF/beans.xml
@@ -3,5 +3,11 @@
        xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee
        http://xmlns.jcp.org/xml/ns/javaee/beans_1_1.xsd"
        version="1.1" bean-discovery-mode="all">
+    
+    <alternatives>
+        <class>
+            com.github.jmetzz.mock.MockGenerator
+        </class>
+    </alternatives>
 
 </beans>
